+++ src/cook-web/src/app/c/[[...id]]/Components/Pay/PayModalM.tsx
@@ -17,7 +17,7 @@ import {
   DialogHeader,
   DialogTitle,
 } from '@/components/ui/dialog';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
+import { RadioGroup, RadioGroupItem } from '@/components/ui/RadioGroup';
 
 import {
   PAY_CHANNEL_WECHAT_JSAPI,
