+++ src/cook-web/src/app/c/[[...id]]/Components/Settings/SettingRadioElement.tsx
@@ -3,7 +3,7 @@ import styles from './SettingRadioElement.module.scss';
 import { useEffect, memo, useState } from 'react';
 
 import { Label } from '@/components/ui/label';
-import { RadioGroup, RadioGroupItem } from '@/components/ui/radio-group';
+import { RadioGroup, RadioGroupItem } from '@/components/ui/RadioGroup';
 
 import { cn } from '@/lib/utils';
