+++ src/cook-web/src/components/render-ui/index.tsx
@@ -32,7 +32,7 @@ import {
   AlertDialogFooter,
   AlertDialogHeader,
   AlertDialogTitle,
-} from '../ui/alert-dialog';
+} from '../ui/AlertDialog';
 import { useTranslation } from 'react-i18next';
 import { memo } from 'react';
 import Empty from './empty';
